Fix integer overflow bug

+++ astra-sim/system/Sys.cc
@@ -1147,7 +1147,7 @@ int Sys::break_dimension(int model_parallel_npu_group) {
   return -1;
 }
 
-int Sys::determine_chunk_size(uint64_t size, ComType type) {
+uint64_t Sys::determine_chunk_size(uint64_t size, ComType type) {
   uint64_t chunk_size = size / preferred_dataset_splits;
   return chunk_size;
 }
